+++ src/main/java/org/trebol/jpa/services/conversion/impl/SalesConverterServiceImpl.java
@@ -42,7 +42,9 @@ import org.trebol.jpa.entities.SellDetail;
 import org.trebol.jpa.entities.Shipper;
 import org.trebol.jpa.repositories.AddressesRepository;
 import org.trebol.jpa.repositories.BillingTypesRepository;
+import org.trebol.jpa.repositories.PaymentTypesRepository;
 import org.trebol.jpa.repositories.ProductsRepository;
+import org.trebol.jpa.repositories.SellStatusesRepository;
 import org.trebol.jpa.repositories.ShippersRepository;
 import org.trebol.jpa.services.conversion.AddressesConverterService;
 import org.trebol.jpa.services.conversion.BillingCompaniesConverterService;
@@ -59,6 +61,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static org.trebol.config.Constants.BILLING_TYPE_ENTERPRISE;
+import static org.trebol.config.Constants.SELL_STATUS_PENDING;
 
 @Transactional
 @Service
@@ -74,6 +77,8 @@ public class SalesConverterServiceImpl
   private final ProductsRepository productsRepository;
   private final ShippersRepository shippersRepository;
   private final AddressesRepository addressesRepository;
+  private final PaymentTypesRepository paymentTypesRepository;
+  private final SellStatusesRepository sellStatusesRepository;
   private final AddressesConverterService addressesConverterService;
   static final double TAX_PERCENT = 0.19; // TODO refactor into a "tax service" of sorts
   static final String UNEXISTING_BILLING_TYPE = "Specified billing type does not exist";
@@ -90,6 +95,8 @@ public class SalesConverterServiceImpl
     ProductsRepository productsRepository,
     ShippersRepository shippersRepository,
     AddressesRepository addressesRepository,
+    PaymentTypesRepository paymentTypesRepository,
+    SellStatusesRepository sellStatusesRepository,
     AddressesConverterService addressesConverterService
   ) {
     this.customersCrudService = customersCrudService;
@@ -102,6 +109,8 @@ public class SalesConverterServiceImpl
     this.productsRepository = productsRepository;
     this.shippersRepository = shippersRepository;
     this.addressesRepository = addressesRepository;
+    this.paymentTypesRepository = paymentTypesRepository;
+    this.sellStatusesRepository = sellStatusesRepository;
     this.addressesConverterService = addressesConverterService;
   }
 
@@ -154,6 +163,9 @@ public class SalesConverterServiceImpl
     if (model.getDate() != null) {
       target.setDate(model.getDate());
     }
+    sellStatusesRepository.findByName(SELL_STATUS_PENDING)
+      .ifPresent(target::setStatus);
+    this.convertPaymentTypeInformationForEntity(model, target);
     this.convertCustomerInformationForEntity(model, target);
     this.convertBillingInformationForEntity(model, target);
     this.convertShippingInformationForEntity(model, target);
@@ -191,10 +203,12 @@ public class SalesConverterServiceImpl
         throw new BadInputException("Unexisting product in sell details");
       }
       Product product = productByBarcode.get();
+      String description = detail.getUnits() + "x " + product.getName();
       return SellDetail.builder()
         .units(detail.getUnits())
         .product(product)
         .unitValue(product.getPrice())
+        .description(description)
         .build();
     } catch (BadInputException exc) {
       throw new RuntimeException(exc.getMessage(), exc);
@@ -204,6 +218,19 @@ public class SalesConverterServiceImpl
   @Override
   public Sell applyChangesToExistingEntity(SellPojo source, Sell target) {
     throw new UnsupportedOperationException("This method is deprecated");
+  }
+
+  private void convertPaymentTypeInformationForEntity(SellPojo model, Sell target) throws BadInputException {
+    String paymentType = model.getPaymentType();
+    if (!StringUtils.isBlank(paymentType)) {
+      paymentTypesRepository.findByName(paymentType)
+        .ifPresentOrElse(target::setPaymentType,
+          () -> {
+            throw new RuntimeException("The payment type does not exist");
+          });
+    } else {
+      throw new BadInputException("A payment type has to be included");
+    }
   }
 
   /**
@@ -245,14 +272,14 @@ public class SalesConverterServiceImpl
       } else {
         target.setBillingCompany(existingCompany.get());
       }
-    }
-    AddressPojo pojoBillingAddress = model.getBillingAddress();
-    Optional<Address> existingBillingAddress = this.findAddress(pojoBillingAddress);
-    if (existingBillingAddress.isEmpty()) {
-      Address address = addressesConverterService.convertToNewEntity(pojoBillingAddress);
-      target.setBillingAddress(address);
-    } else {
-      target.setBillingAddress(existingBillingAddress.get());
+      AddressPojo pojoBillingAddress = model.getBillingAddress();
+      Optional<Address> existingBillingAddress = this.findAddress(pojoBillingAddress);
+      if (existingBillingAddress.isEmpty()) {
+        Address address = addressesConverterService.convertToNewEntity(pojoBillingAddress);
+        target.setBillingAddress(address);
+      } else {
+        target.setBillingAddress(existingBillingAddress.get());
+      }
     }
   }