+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.1] - 2023-06-27
+
+### Fixed
+
+- Raise an exception when no payment type is specified
+- Unintended errors during checkout
+  - Description of each sell detail would not be included
+  - New sales would not have an initial status or registered payment type
+  - Billing address was required for non-enterprise customers>
+
 ## [v0.2.0] - 2023-05-26
 
 ### Added